fix: wordpress

+++ src/wordpress/tools/entrypoint.sh
@@ -7,16 +7,18 @@ err() {
 }
 
 setup_wordpress() {
- cd ${WP_PATH} && wp config create --dbname=${MARIADB_NAME} --dbuser=${MARIADB_USER} --dbpass=${MARIADB_PASSWORD} --dbhost=localhost --path=${WP_PATH} --allow-root || return 1
+ if [[ -e /var/www/html/wordpress/wp-config.php ]]; then
+   return 0
+ fi
+ cd ${WP_PATH} && wp config create --dbname=${MARIADB_NAME} --dbuser=${MARIADB_USER} --dbpass=${MARIADB_PASSWORD} --dbhost=mariadb --path=${WP_PATH} --allow-root || return 1
  wp core install --allow-root --path=${WP_PATH} --title=${WP_TITLE} --admin_user=${WP_ADMIN_NAME} --admin_password=${WP_ADMIN_PASSWORD} --admin_email=${WP_ADMIN_EMAIL} --url=${DOMAIN_NAME} || return 1
  wp user create ${WP_USER_NAME} ${WP_USER_EMAIL} --role=subscriber --user_pass=${WP_USER_PASSWORD} --allow-root --path=${WP_PATH} || return 1
 return 0
 }
 
 main() {
-  service php8.2-fpm start || { err "failed to start php-fpm"; return 1; }
   setup_wordpress || { err "failed to setup wordpress"; return 1; }
-  php-fpm8.2 -F
+  php-fpm8.2 -F || { err "failed to start php-fpm"; return 1; }
   return 0
 }
